+++ src-vm/obj.zig
@@ -32,6 +32,17 @@ pub const ObjString = struct {
         };
     }
 
+    /// Inits the ObjString from 2 slices, concatenated.
+    pub fn init_dynamic_from_2(self: *Self, allocator: std.mem.Allocator, bytes_1: []const u8, bytes_2: []const u8) !void {
+        const slices: [2][]const u8 = .{ bytes_1, bytes_2 };
+        const new_str = try std.mem.concat(allocator, u8, &slices);
+
+        self.* = .{
+            .base = .{ .t = ObjType.String },
+            .bytes = .{ .heap = try allocator.dupe(u8, new_str) },
+        };
+    }
+
     pub fn deinit(self: *Self, allocator: std.mem.Allocator) void {
         switch (self.bytes) {
             .heap => |h| {